% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{parallelInnerProduct}
\alias{parallelInnerProduct}
\title{Parallel Inner Product}
\usage{
parallelInnerProduct(lhs, rhs)
}
\arguments{
\item{lhs, rhs}{Numeric vectors.}
}
\description{
Compute the inner product of two numeric vectors.
}

