\name{rcpp-parallel-package}
\alias{rcpp-parallel-package}
\alias{rcpp-parallel}
\docType{package}
\title{
\packageTitle{rcpp-parallel}
}
\description{
\packageDescription{rcpp-parallel}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{rcpp-parallel}
\packageIndices{rcpp-parallel}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{rcpp-parallel}

Maintainer: \packageMaintainer{rcpp-parallel}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
